+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - [BREAKING] Make Falcon DSA deterministic (#436).
 - `Smt.insert_inner_nodes` does not store empty subtrees ([#452](https://github.com/0xMiden/crypto/pull/452)).
 
+## 0.15.6 (2025-07-15)
+
+- Added conversions and seralization for `PartialSmt` ([#451](https://github.com/0xMiden/crypto/pull/451/), [#453](https://github.com/0xMiden/crypto/pull/453/)).
+
 ## 0.15.5 (2025-07-10)
 
 - Added `empty()` and `is_empty()` methods to `Word`.